tech: refactor compatbility layer initialization and move things around
