+++ src/backend/main.ts
@@ -53,16 +53,12 @@ import {
   getStoreName,
   isEpicServiceOffline,
   handleExit,
-  checkRosettaInstall,
   openUrlOrFile,
   resetApp,
-  setGPTKDefaultOnMacOS,
   showAboutWindow,
   showItemInFolder,
   wait,
-  getShellPath,
-  checkWineBeforeLaunch,
-  downloadDefaultWine
+  getShellPath
 } from './utils'
 import {
   configPath,
@@ -82,7 +78,6 @@ import {
   isCLIFullscreen,
   isCLINoGui,
   isFlatpak,
-  isMac,
   isSteamDeckGameMode,
   onboardLocalStore,
   publicDir,
@@ -151,6 +146,13 @@ import { legendarySetup } from 'backend/storeManagers/legendary/setup'
 import * as Sentry from '@sentry/electron'
 import { DEV_PORTAL_URL, devSentryDsn, prodSentryDsn } from 'common/constants'
 import { getHpOverlay, initOverlay } from './overlay'
+
+import { initExtension } from './extension/importer'
+import { hpApi } from './utils/hyperplay_api'
+import {
+  initializeCompatibilityLayer,
+  checkWineBeforeLaunch
+} from './utils/compatibility_layers'
 
 /*
  * INSERT OTHER IPC HANDLERS HERE
@@ -649,20 +651,10 @@ if (!gotTheLock) {
     }
 
     // Will download Wine if none was found
-    const availableWine = (await GlobalConfig.get().getAlternativeWine()) || []
-    const toolkitDownloaded = availableWine.some(
-      (wine) => wine.type === 'toolkit'
-    )
-    const shouldDownloadWine =
-      !availableWine.length || (isMac && !toolkitDownloaded)
-
-    Promise.all([
-      DXVK.getLatest(),
-      Winetricks.download(),
-      shouldDownloadWine ? downloadDefaultWine() : null,
-      isMac && checkRosettaInstall(),
-      isMac && !shouldDownloadWine && setGPTKDefaultOnMacOS()
-    ])
+    // Call the initialization function at the appropriate place
+    if (!isWindows) {
+      initializeCompatibilityLayer()
+    }
 
     // set initial zoom level after a moment, if set in sync the value stays as 1
     setTimeout(() => {
@@ -2106,5 +2098,3 @@ ipcMain.handle('getHyperPlayListings', async () => {
  */
 
 import './storeManagers/legendary/eos_overlay/ipc_handler'
-import { initExtension } from './extension/importer'
-import { hpApi } from './utils/hyperplay_api'
