+++ src/backend/utils/compatibility_layers.ts
@@ -2,20 +2,59 @@ import { GlobalConfig } from 'backend/config'
 import {
   configPath,
   getSteamLibraries,
+  icon,
+  isLinux,
   isMac,
+  isWindows,
   toolsPath,
   userHome
 } from 'backend/constants'
-import { logError, LogPrefix, logInfo } from 'backend/logger/logger'
-import { execAsync } from 'backend/utils'
+import {
+  logError,
+  LogPrefix,
+  logInfo,
+  logsDisabled
+} from 'backend/logger/logger'
+import {
+  checkRosettaInstall,
+  execAsync,
+  isMacSonomaOrHigher
+} from 'backend/utils'
 import { execSync } from 'child_process'
-import { WineInstallation } from 'common/types'
-import { existsSync, mkdirSync, readFileSync, readdirSync } from 'graceful-fs'
+import {
+  GameSettings,
+  ProgressInfo,
+  State,
+  WineInstallation
+} from 'common/types'
+import {
+  appendFileSync,
+  existsSync,
+  mkdirSync,
+  readFileSync,
+  readdirSync
+} from 'graceful-fs'
 import { homedir } from 'os'
 import { dirname, join } from 'path'
 import { PlistObject, parse as plistParse } from 'plist'
 import LaunchCommand from '../storeManagers/legendary/commands/launch'
 import { NonEmptyString, Path } from '../storeManagers/legendary/commands/base'
+import { GameConfig } from 'backend/game_config'
+import { validWine } from 'backend/launcher'
+import { sendFrontendMessage } from 'backend/main_window'
+import { DXVK, Winetricks } from 'backend/tools'
+import {
+  installWineVersion,
+  updateWineVersionInfos,
+  wineDownloaderInfoStore
+} from 'backend/wine/manager/utils'
+import { dialog } from 'electron'
+import i18next from 'i18next'
+import {
+  createAbortController,
+  deleteAbortController
+} from './aborthandler/aborthandler'
+import { notify } from 'backend/dialog/dialog'
 
 /**
  * Loads the default wine installation path and version.
@@ -525,4 +564,298 @@ export function getWineFlagsArray(
     else commandArray.push(key, value)
   }
   return commandArray
+}
+
+export async function initializeCompatibilityLayer() {
+  // Fetch available Wine versions on the system
+  const availableWine = (await GlobalConfig.get().getAlternativeWine()) || []
+
+  // Determine if the toolkit has been downloaded on macOS
+  const toolkitDownloaded = availableWine.some(
+    (wine) => wine.type === 'toolkit'
+  )
+
+  // Determine if we need to download a new Wine version
+  const shouldDownloadWine =
+    !availableWine.length || (isMac && !toolkitDownloaded)
+
+  // Build an array of promises for initialization tasks
+  const initializationTasks: Array<Promise<WineInstallation | void | null>> = [
+    DXVK.getLatest(),
+    Winetricks.download()
+  ]
+
+  if (shouldDownloadWine) {
+    initializationTasks.push(downloadDefaultWine())
+  }
+
+  if (isMac) {
+    initializationTasks.push(checkRosettaInstall())
+    initializationTasks.push(setGPTKDefaultOnMacOS())
+  }
+
+  try {
+    await Promise.all(initializationTasks)
+  } catch (error) {
+    logError(
+      ['Error during compatibility layer initialization', error],
+      LogPrefix.Backend
+    )
+  }
+}
+
+export async function downloadDefaultWine() {
+  if (isWindows) return null
+
+  try {
+    // Refresh wine list
+    await updateWineVersionInfos(true)
+
+    // Get list of available wine versions
+    const availableWine = wineDownloaderInfoStore.get('wine-releases', [])
+
+    // Determine the correct wine version based on platform and compatibility
+    const release = availableWine.find(async (version) => {
+      if (isLinux) {
+        return (
+          version.type === 'Wine-GE' &&
+          version.version.includes('Wine-GE-Proton')
+        )
+      } else if (isMac) {
+        const isGPTKCompatible = await isMacSonomaOrHigher()
+        return isGPTKCompatible
+          ? version.type === 'Game-Porting-Toolkit'
+          : version.type === 'Wine-Crossover'
+      }
+      return false
+    })
+
+    if (!release) {
+      logError('Could not find default wine version', LogPrefix.Backend)
+      return null
+    }
+
+    // Notify user and start download
+    notify({
+      title: i18next.t(
+        'notification.wine-download.title',
+        'Compatibility Layer'
+      ),
+      body: i18next.t(
+        'notification.wine-download.message',
+        'Setting up the default compatibility layer'
+      )
+    })
+
+    const onProgress = (state: State, progress?: ProgressInfo) => {
+      sendFrontendMessage(`progressOfWineManager${release.version}`, {
+        state,
+        progress
+      })
+    }
+
+    const abortController = createAbortController(release.version)
+    const result = await installWineVersion(
+      release,
+      onProgress,
+      abortController.signal
+    )
+
+    deleteAbortController(release.version)
+
+    if (result !== 'success') {
+      return null
+    }
+
+    // Update the game config to use the downloaded wine
+    const wineList = await GlobalConfig.get().getAlternativeWine()
+    const downloadedWine = wineList[0]
+
+    if (downloadedWine) {
+      logInfo(`Changing wine version to ${downloadedWine.name}`)
+      GlobalConfig.get().setSetting('wineVersion', downloadedWine)
+    }
+
+    return downloadedWine
+  } catch (error) {
+    logError(['Error during wine download process', error], LogPrefix.Backend)
+    notify({
+      title: i18next.t('notification.wine-download-failed.title', 'Error'),
+      body: i18next.t(
+        'notification.wine-download-failed.message',
+        'Failed to setup the default compatibility layer'
+      )
+    })
+
+    return null
+  }
+}
+
+export async function setGPTKDefaultOnMacOS() {
+  const isGPTKCompatible = await isMacSonomaOrHigher()
+  if (!isGPTKCompatible) {
+    return
+  }
+
+  const { wineVersion: defaultWine } = GlobalConfig.get().getSettings()
+
+  const ignoreList = ['crossover', 'toolkit']
+
+  if (
+    ignoreList.includes(defaultWine.type.toLowerCase()) ||
+    defaultWine.name.includes('Toolkit')
+  ) {
+    return
+  }
+
+  const wineList = await GlobalConfig.get().getAlternativeWine()
+  const gptk = wineList.find((wine) => wine.type === 'toolkit')
+
+  if (gptk && existsSync(gptk.bin)) {
+    logInfo(`Changing wine version to ${gptk.name}`)
+    GlobalConfig.get().setSetting('wineVersion', gptk)
+    // update prefix to use the new one as well
+    const installPath = GlobalConfig.get().getSettings().defaultInstallPath
+    const newPrefix = join(installPath, 'Prefixes', 'GPTK')
+    GlobalConfig.get().setSetting('winePrefix', newPrefix)
+  }
+  return
+}
+
+export async function checkWineBeforeLaunch(
+  appName: string,
+  gameSettings: GameSettings,
+  logFileLocation: string
+): Promise<boolean> {
+  const wineIsValid = await validWine(gameSettings.wineVersion)
+
+  const isToolkit = gameSettings.wineVersion.type === 'toolkit'
+  const isGPTKCompatible = await isMacSonomaOrHigher()
+
+  const isValidOnLinux = isLinux && wineIsValid
+  const isValidtoolkitOnMac =
+    isMac && isToolkit && isGPTKCompatible && wineIsValid
+  const isValidWineOnMac = isMac && !isToolkit && wineIsValid
+  const isValidOnMac = isValidtoolkitOnMac || isValidWineOnMac
+
+  if (isValidOnMac || isValidOnLinux) {
+    return true
+  } else {
+    if (!logsDisabled) {
+      logError(
+        `Wine version ${gameSettings.wineVersion.name} is not valid, trying another one.`,
+        LogPrefix.Backend
+      )
+
+      appendFileSync(
+        logFileLocation,
+        `Wine version ${gameSettings.wineVersion.name} is not valid, trying another one.`
+      )
+    }
+
+    // check if the default wine is valid now
+    const { wineVersion: defaultwine } = GlobalConfig.get().getSettings()
+    const defaultWineIsValid = await validWine(defaultwine)
+    if (defaultWineIsValid) {
+      const { response } = await ContinueWithFoundWine(
+        gameSettings.wineVersion.name,
+        defaultwine.name
+      )
+
+      if (response === 0) {
+        logInfo(`Changing wine version to ${defaultwine.name}`)
+        gameSettings.wineVersion = defaultwine
+        GameConfig.get(appName).setSetting('wineVersion', defaultwine)
+        return true
+      } else {
+        logInfo('User canceled the launch', LogPrefix.Backend)
+        return false
+      }
+    } else {
+      const wineList = await GlobalConfig.get().getAlternativeWine()
+
+      // if Linux get the first element, if macOS and isGPTKCompatible is true get one with type 'toolkit', otherwise get the one with type 'wine'
+      const firstFoundWine = wineList.find((wine) => {
+        if (isLinux) {
+          return wine.type === 'wine'
+        } else if (isMac) {
+          return isGPTKCompatible
+            ? wine.type === 'toolkit'
+            : wine.type === 'wine'
+        }
+        return undefined
+      })
+
+      const isValidWine = await validWine(firstFoundWine)
+
+      if (!wineList.length || !firstFoundWine || !isValidWine) {
+        const firstFoundWine = await downloadDefaultWine()
+        if (firstFoundWine) {
+          logInfo(`Changing wine version to ${firstFoundWine.name}`)
+          gameSettings.wineVersion = firstFoundWine
+          GameConfig.get(appName).setSetting('wineVersion', firstFoundWine)
+          return true
+        }
+      }
+
+      if (firstFoundWine && isValidWine) {
+        const { response } = await ContinueWithFoundWine(
+          gameSettings.wineVersion.name,
+          firstFoundWine.name
+        )
+
+        if (response === 0) {
+          logInfo(`Changing wine version to ${firstFoundWine.name}`)
+          gameSettings.wineVersion = firstFoundWine
+          GameConfig.get(appName).setSetting('wineVersion', firstFoundWine)
+          return true
+        } else {
+          logInfo('User canceled the launch', LogPrefix.Backend)
+          return false
+        }
+      }
+    }
+  }
+  return false
+}
+
+async function ContinueWithFoundWine(
+  selectedWine: string,
+  foundWine: string
+): Promise<{ response: number }> {
+  const isGPTK = selectedWine.toLowerCase().includes('toolkit')
+  const isGPTKCompatible = await isMacSonomaOrHigher()
+
+  if (isMac && isGPTK && !isGPTKCompatible) {
+    const { response } = await dialog.showMessageBox({
+      title: i18next.t(
+        'box.warning.wine-change.title-gptk',
+        'Game Porting Toolkit Not Compatible '
+      ),
+      message: i18next.t('box.warning.wine-change.message-gptk', {
+        defaultValue:
+          'To be able to run games using the Apple Gaming porting toolkit you need to upgrade your macOS to 14 (Sonoma) or higher. {{newline}} We found Wine on your system, do you want to continue launching using {{foundWine}} ?',
+        newline: '\n',
+        foundWine: foundWine
+      }),
+      buttons: [i18next.t('box.yes'), i18next.t('box.no')],
+      icon: icon
+    })
+    return { response }
+  }
+
+  const { response } = await dialog.showMessageBox({
+    title: i18next.t('box.warning.wine-change.title', 'Wine not found!'),
+    message: i18next.t('box.warning.wine-change.message', {
+      defaultValue:
+        'We could not find the selected wine version to launch this title ({{selectedWine}}). {{newline}} We found another one, do you want to continue launching using {{foundWine}} ?',
+      newline: '\n',
+      selectedWine: selectedWine,
+      foundWine: foundWine
+    }),
+    buttons: [i18next.t('box.yes'), i18next.t('box.no')],
+    icon: icon
+  })
+
+  return { response }
 }
